+++ lib/sucker_punch/railtie.rb
@@ -1,0 +1,7 @@
+module SuckerPunch
+  class Railtie < ::Rails::Railtie
+    initializer "sucker_punch.logger" do
+      SuckerPunch.logger = Rails.logger
+    end
+  end
+end