+++ spec/sucker_punch_spec.rb
@@ -38,4 +38,15 @@ describe SuckerPunch do
     end
 
   end
+
+  describe 'logger' do
+    it "delegates get to Celluloid's logger" do
+      SuckerPunch.logger.should == Celluloid.logger
+    end
+
+    it "delegates set to Celluloid's logger" do
+      Celluloid.should_receive(:logger=)
+      SuckerPunch.logger = nil
+    end
+  end
 end